WotW Update 8/29
-Update WotW Core Localization
-Updated Score Modlet

+++ Mods/0-SCore/Features/FireV2/Scripts/FireHandler.cs
@@ -487,6 +487,8 @@ public class FireHandler : IFireHandler
     {
         var pooledExpandableMemoryStream =
             (PooledExpandableMemoryStream)threadInfo.parameter;
+        if (!Directory.Exists(GameIO.GetSaveGameDir())) return -1;
+
         var text = $"{GameIO.GetSaveGameDir()}/{SaveFile}";
         if (File.Exists(text))
             File.Copy(text, $"{GameIO.GetSaveGameDir()}/{SaveFile}.bak", true);
